+++ vega-lite/src/main/kotlin/dev/aster/vegalite/UnitView.kt
@@ -256,6 +256,7 @@ internal class UnitView(
           height = spec.height,
           params = spec.params,
           projection = spec.projection,
+          viewBackground = spec.viewBackground,
         ),
         config,
         name,